feat: improve integration tests
